# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TraceCrypt Unified \u2014 Evidence-led Attribution",
    "description": "Read-only investigative triage. Does not identify persons, establish guilt, guarantee recovery, or execute freezes.",
    "version": "unified-2.0"
  },
  "paths": {
...
    "/api/v1/cases/{case_id}/investigations": {
      "post": {
        "tags": [
          "unified-workspace"
        ],
        "summary": "Investigate",
        "operationId": "investigate_api_v1_cases__case_id__investigations_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "unified-workspace"
        ],
        "summary": "Investigations",
        "operationId": "investigations_api_v1_cases__case_id__investigations_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/investigations/{run_id}": {
      "get": {
        "tags": [
          "unified-workspace"
        ],
        "summary": "Investigation",
        "operationId": "investigation_api_v1_cases__case_id__investigations__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/investigations/{run_id}/report": {
      "get": {
        "tags": [
          "unified-workspace"
        ],
        "summary": "Report",
        "operationId": "report_api_v1_cases__case_id__investigations__run_id__report_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/investigations/{run_id}/export": {
      "get": {
        "tags": [
          "unified-workspace"
        ],
        "summary": "Export",
        "operationId": "export_api_v1_cases__case_id__investigations__run_id__export_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/investigations/{run_id}/request-draft": {
      "post": {
        "tags": [
          "unified-workspace"
        ],
        "summary": "Request Draft",
        "operationId": "request_draft_api_v1_cases__case_id__investigations__run_id__request_draft_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedDraftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cases/{case_id}/watches/{watch_id}/poll": {
      "post": {
        "tags": [
          "unified-workspace"
        ],
        "summary": "Poll",
        "operationId": "poll_api_v1_cases__case_id__watches__watch_id__poll_post",
        "parameters": [
          {
            "name": "watch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Watch Id"
            }
          },
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workspace/demo/{preset_id}": {
      "get": {
        "tags": [
          "local-demo-only"
        ],
        "summary": "Demo Detail",
        "operationId": "demo_detail_api_v1_workspace_demo__preset_id__get",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InvestigationRequest": {
        "properties": {
          "network_key": {
            "type": "string",
            "title": "Network Key"
          },
          "address": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Address"
          },
          "asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "token_contract": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Contract"
          },
          "seed_event_reference": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed Event Reference"
          },
          "mode": {
            "$ref": "#/components/schemas/SeedMode",
            "default": "incident"
          },
          "analysis_cutoff": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Cutoff"
          }
        },
        "type": "object",
        "required": [
          "network_key",
          "address"
        ],
        "title": "InvestigationRequest"
      },
...
      "SavedDraftRequest": {
        "properties": {
          "request_kind": {
            "type": "string",
            "title": "Request Kind"
          },
          "target_address": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Target Address"
          },
          "requesting_agency": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Requesting Agency"
          },
          "requesting_officer": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Requesting Officer"
          },
          "alleged_incident_summary": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Alleged Incident Summary"
          },
          "legal_authority_reference": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Authority Reference"
          },
          "exchange_account_identifiers": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange Account Identifiers"
          }
        },
        "type": "object",
        "required": [
          "request_kind",
          "target_address",
          "requesting_agency",
          "requesting_officer",
          "alleged_incident_summary"
        ],
        "title": "SavedDraftRequest"
      },
...
      "SeedMode": {
        "type": "string",
        "enum": [
          "incident",
          "address_discovery"
        ],
        "title": "SeedMode"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}